requirements.txt
torch
scikit-learn
numpy
tqdm
faiss-cpu
pyyaml
